refactor: use composite action instead of reusable workflows
Unify the monorepo and the non-monorepo starter actions by making both
use the composite action instead of the reusable workflows (which we
cannot use for the monorepo workflow).

The composite action has feature paritiy with the previous approach and
does not depend on `yq` container, making it faster and thus the
prefered choice for our non-monorepo action as well.